Revert "issues-map: relocate domain implementation to cnos.issues package boundary"
This reverts commit 88156120cfb841fb21cf95c59268643eefc5098e.
