+++ src/go/internal/issuesmap/templates/board.template.html
